feat(database): graduate RTDB CDD surface
