+++ packages/cli/src/serve/worker/index.ts
@@ -202,12 +202,17 @@ export {
   rtdbOnValue,
   rtdbOnChildAdded,
   rtdbOnChildChanged,
+  rtdbOnChildMoved,
+  rtdbOnChildRemoved,
   rtdbOnDisconnect,
   rtdbGoOffline,
   rtdbGoOnline,
   RtdbOnDisconnect,
   rtdbOff,
   rtdbServerTimestamp,
+  rtdbSetPriority,
+  rtdbSetWithPriority,
+  rtdbRunTransaction,
   rtdbConnectDatabaseEmulator,
   adminReadRtdbState,
   adminSetRtdbValue,
