+++ packages/cli/test/serve/worker/client-surface.test.ts
@@ -38,8 +38,10 @@ const EXPECTED_VALUE_EXPORTS: readonly string[] = [
   'onAuthStateChanged', 'onIdTokenChanged', 'onSnapshot', 'or', 'orderBy', 'query',
   'ref', 'relayWorkerOp', 'relayWorkerSub', 'resetAll', 'restorePortSession', 'rtdbChild',
   'rtdbConnectDatabaseEmulator', 'rtdbGet', 'rtdbGetDatabase', 'rtdbGoOffline',
-  'rtdbGoOnline', 'rtdbOff', 'rtdbOnChildAdded', 'rtdbOnChildChanged', 'rtdbOnDisconnect', 'rtdbOnValue',
-  'rtdbPush', 'rtdbRef', 'rtdbRemove', 'rtdbServerTimestamp', 'rtdbSet', 'rtdbUpdate',
+  'rtdbGoOnline', 'rtdbOff', 'rtdbOnChildAdded', 'rtdbOnChildChanged', 'rtdbOnChildMoved',
+  'rtdbOnChildRemoved', 'rtdbOnDisconnect', 'rtdbOnValue', 'rtdbPush', 'rtdbRef',
+  'rtdbRemove', 'rtdbRunTransaction', 'rtdbServerTimestamp', 'rtdbSet', 'rtdbSetPriority',
+  'rtdbSetWithPriority', 'rtdbUpdate',
   'runTransaction', 'saveWorkerBranch', 'serverTimestamp', 'setDatabaseRules', 'setDoc',
   'setFirestoreRules', 'setLens', 'setOpIssuer', 'setPersistence',
   'setProviderConfig', 'setRules', 'signInAnonymously', 'signInWithCredential',
